BJS2-27264 - перенесён пакет dto, добавлены интерфейсы сервисов
