--- src/main/java/faang/school/postservice/dto/correcter/SpellCheckResponse.java
@@ -1,12 +1,0 @@
-package faang.school.postservice.dto.correcter;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class SpellCheckResponse {
-    private String corrected;
-}
