+++ src/main/java/faang/school/postservice/model/dto/correcter/SpellCheckResponse.java
@@ -1,4 +1,4 @@
-package faang.school.postservice.dto.correcter;
+package faang.school.postservice.model.dto.correcter;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class TextGearsResponse<T> {
-    private boolean status;
-    private T response;
+public class SpellCheckResponse {
+    private String corrected;
 }
