+++ src/main/java/faang/school/postservice/model/dto/correcter/TextGearsResponse.java
@@ -1,0 +1,13 @@
+package faang.school.postservice.model.dto.correcter;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class TextGearsResponse<T> {
+    private boolean status;
+    private T response;
+}
