+++ src/main/java/faang/school/postservice/service/BatchProcessService.java
@@ -1,0 +1,11 @@
+package faang.school.postservice.service;
+
+import faang.school.postservice.model.Post;
+
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
+
+public interface BatchProcessService {
+
+    CompletableFuture<Void> processBatch(List<Post> postsBatch);
+}
