+++ src/main/java/faang/school/postservice/service/SpellCheckService.java
@@ -1,0 +1,8 @@
+package faang.school.postservice.service;
+
+public interface SpellCheckService {
+
+    String autoCorrect(String content, String language);
+
+    String detectLanguage(String text);
+}
